pyproject.toml
[tool.poetry]
name = "suraksha"
version = "1.0.0"
description = "SuRक्षा provides smart security surveillance solutions such as face recognition, object detection, and more."
authors = ["ChiragAgg5k <110609663+ChiragAgg5k@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ultralytics = "^8.1.11"
torch = "^2.2.0"
torchvision = "^0.17.0"
torchaudio = "^2.2.0"
flask = "^3.0.2"
flask-mail = "^0.9.1"
pyrebase4 = "^4.7.1"
black = "^24.4.0"
pytest = "^8.1.1"
firebase = "^4.0.1"
transformers = "^4.40.0"
setuptools = "^72.1.0"
python-dotenv = "^1.0.1"
krutrim-cloud = "^0.6.1"
firebase-admin = "^6.5.0"
openai = "^1.52.0"
numpy = ">=2.1.0"
opencv-python = "^4.11.0.86"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
